Return a list of string classifications from column_detection eval
Rather than returning a list of nubmers representing the classifications,
return a list of strings that show the classifications. This lines up with
the previously used fake_column_detector which returns string labels.

Also modify frontend to call the column detector with all the lines on
a page, rather than once per line. This is more efficient.

The frontend can now use either the fake column detector or the real one.
The fake version is used during training when we need completely correct
column segmentation.