Stop reading image files when training line classifier
Source the input data from data_loader and call frontend directly.

Previously, we used a script to call the frontend, and write the features to image
files. Then training read those files from disk. Now that we can use frontend as a library
simplyfiy the training process and just call the frontend, alliviating the need to write features
to disk.