Use a background image when computing features
This background image is a render of the pdf page. The intention is to
still mask off the characters, but add extra picture data to the features.
The hope is this will improve the separation between figure text and table
classification.

The features for text only ares will not change significantly with this change.

In the future we will not read images from disk, but rather directly from the pdf.