Fix viterbi scoring
Previously viterbi was calculating the reference based on how the line classifier was splitting the columns. Now we calculate columns for the reference and hypothesis seperately. The reference uses the labeld boxes to detect columns, and the hypothesis uses the column model.

Now since the number of hypothesis lines can be different than the number of reference lines, we align the hypothesis and reference on a per page basis.