Allow configuring how to iterpret manual labels
Previously, we assumed labels were in the form of
"(Single|Double)Column-<line_label>". If a training set
is all single column, it doesn't make much sense to
mark everything as single. The function in config.py
can be customized to not split the string and return
'SingleColumn' all the time.

It could also be customized if the labels are in another format.