Stop looking for files with the page number specified twice
Previously, csv files were of the form <hash>_<page_num>_<page_num>.csv

This was a hold over from IQC that for some reason wrote files with the page number
twice. This makes no sense, so lets not support this anymore.

Files in the training dir will be  <hash>_<page_num>.<extension>