Move the code for docker
Tabio now runs on docker

Moved api file for tabio docker

Fix import errors with Tabio

Tabio when running from the FastAPI was throwing import errors.
We found the fix for this was to change the imports to tabio. which then
allowed the API to properly import Tabio.

Removed pdf_from_page_data in data_loader.py. Unused function.